+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 CHANGELOG
 ===
+## v0.3.0
+* new pipe - `ReplaceRegExp`
+* update pipe - `Replace` now receives array of strings as `search` parameter for multiple replacements
 ## v0.2.0
-* New pipe - `Replace`
-* New filter - `Unique`
-* New feature `storage` with dependencies `pickleDb` and `md5`
+* new pipe - `Replace`
+* new filter - `Unique`
+* new feature `Storage` with dependencies `pickledb` and `md5`
 ## v0.1.0
 * Initial version